[강운덕] [PROFILER-31] 중복로직을 traceContext로 병합함.
git-svn-id: http://svn.bds.nhncorp.com/pe/hippo-tomcat-profiler/trunk@4101 84d0f5b1-2673-498c-a247-62c4ff18d310

+++ src/main/java/com/nhn/pinpoint/profiler/monitor/metric/LongAdderHistogram.java
@@ -33,7 +33,7 @@ public class LongAdderHistogram implements Histogram {
 
     public void addResponseTime(int millis) {
         HistogramSchema schema = serviceType.getHistogramSchema();
-        final HistogramSlot histogramSlot = schema.findHistogramSlot((int) millis);
+        final HistogramSlot histogramSlot = schema.findHistogramSlot(millis);
         final SlotType slotType = histogramSlot.getSlotType();
         switch (slotType) {
             case FAST:
